Refactor: Remove home screen navigation from onboarding
