--- api/onboarding/src/main/java/com/giraffe/onboarding/OnBoardingApi.kt
@@ -1,8 +1,0 @@
-package com.giraffe.onboarding
-
-import androidx.compose.runtime.Composable
-
-interface OnBoardingApi {
-    @Composable
-    fun OnBoardingContainer()
-}
