+++ app/src/main/java/com/giraffe/cineverseapp/di/featureApiModule.kt
@@ -10,10 +10,10 @@ import com.giraffe.home.HomeApi
 import com.giraffe.home.navigation.HomeApiImp
 import com.giraffe.match.MatchApi
 import com.giraffe.match.navigation.MatchApiImp
-import com.giraffe.onboarding.OnBoardingApi
 import com.giraffe.onboarding.nav.OnBoardingApiImp
 import com.giraffe.profile.ProfileApi
 import com.giraffe.profile.navigation.ProfileApiImp
+import com.giraffee.onboarding.OnBoardingApi
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
