+++ presentation/onboarding/build.gradle.kts
@@ -41,7 +41,6 @@ dependencies {
     implementation(project(":domain:user"))
     implementation(project(":api:onboarding"))
     implementation(project(":api:authentication"))
-    implementation(project(":api:home"))
 
     implementation(libs.kotlinx.datetime)
     androidTestImplementation(platform(libs.androidx.compose.bom))
