+++ presentation/onboarding/src/main/java/com/giraffe/onboarding/nav/LoginRoute.kt
@@ -1,4 +1,4 @@
-package com.giraffe.onboarding.nav.routes
+package com.giraffe.onboarding.nav
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
