+++ presentation/onboarding/src/main/java/com/giraffe/onboarding/nav/OnBoardingApiImp.kt
@@ -3,13 +3,11 @@ package com.giraffe.onboarding.nav
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.rememberNavController
 import com.giraffe.authentication.AuthenticationApi
-import com.giraffe.home.HomeApi
-import com.giraffe.onboarding.OnBoardingApi
+import com.giraffee.onboarding.OnBoardingApi
 import jakarta.inject.Inject
 
 class OnBoardingApiImp @Inject constructor(
     private val authenticationApi: AuthenticationApi,
-    private val homeApi: HomeApi
 ) : OnBoardingApi {
 
     @Composable
@@ -17,7 +15,6 @@ class OnBoardingApiImp @Inject constructor(
         val navController = rememberNavController()
         OnBoardingNavGraph(
             navController = navController,
-            homeApi = homeApi,
             authenticationApi = authenticationApi
         )
     }
