+++ presentation/onboarding/src/main/java/com/giraffe/onboarding/nav/OnboardingNavGraph.kt
@@ -4,16 +4,12 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import com.giraffe.authentication.AuthenticationApi
-import com.giraffe.home.HomeApi
-import com.giraffe.onboarding.nav.routes.homeRoute
-import com.giraffe.onboarding.nav.routes.loginRoute
 import com.giraffe.onboarding.screen.OnBoardingRoute
 import com.giraffe.onboarding.screen.onBoardingRoute
 
 @Composable
 internal fun OnBoardingNavGraph(
     navController: NavHostController,
-    homeApi: HomeApi,
     authenticationApi: AuthenticationApi
 ) {
     NavHost(
@@ -23,7 +19,5 @@ internal fun OnBoardingNavGraph(
         onBoardingRoute(navController)
 
         loginRoute(authenticationApi)
-
-        homeRoute(homeApi)
     }
 }
