+++ presentation/onboarding/src/main/java/com/giraffe/onboarding/screen/OnboardingEffect.kt
@@ -4,6 +4,5 @@ import androidx.annotation.StringRes
 
 sealed class OnboardingEffect {
     object NavigateToLogin : OnboardingEffect()
-    object NavigateToHome : OnboardingEffect()
     data class ShowError(@param:StringRes val messageRes: Int) : OnboardingEffect()
 }
