+++ presentation/onboarding/src/main/java/com/giraffe/onboarding/screen/OnboardingInteractionListener.kt
@@ -4,6 +4,5 @@ package com.giraffe.onboarding.screen
 interface OnboardingInteractionListener {
     fun checkIfFirstTime()
     fun markOnboardingComplete()
-    fun navigateToHomeScreen()
     fun navigateToLoginScreen()
 }
