+++ presentation/onboarding/src/main/java/com/giraffe/onboarding/screen/OnboardingViewModel.kt
@@ -65,10 +65,6 @@ class OnboardingViewModel @Inject constructor(
         }
     }
 
-    override fun navigateToHomeScreen() {
-        sendEffect(OnboardingEffect.NavigateToHome)
-    }
-
     override fun navigateToLoginScreen() {
         sendEffect(OnboardingEffect.NavigateToLogin)
     }
